+++ src/icclim/_core/generic/functions.py
@@ -21,6 +21,7 @@ import operator
 from collections.abc import Callable
 from functools import partial
 from typing import TYPE_CHECKING, Any
+import warnings
 from warnings import warn
 
 import xarray as xr
@@ -137,7 +138,7 @@ def count_occurrences(
         result.attrs[UNITS_KEY] = "%"
         return result
     freq = check_freq(climate_vars[0].studied_data, dim="time")
-    return to_agg_units(result, climate_vars[0].studied_data, "count", deffreq=freq)
+    return _safe_to_agg_units(result, climate_vars[0].studied_data, "count", deffreq=freq)
 
 
 def max_consecutive_occurrence(
@@ -209,7 +210,7 @@ def max_consecutive_occurrence(
     else:
         result = resampled.max(dim="time")
     freq = check_freq(climate_vars[0].studied_data, dim="time")
-    return to_agg_units(result, climate_vars[0].studied_data, "count", deffreq=freq)
+    return _safe_to_agg_units(result, climate_vars[0].studied_data, "count", deffreq=freq)
 
 
 def sum_of_spell_lengths(
@@ -254,7 +255,7 @@ def sum_of_spell_lengths(
     cropped_rle = rle.where(rle >= min_spell_length, other=0)
     result = cropped_rle.resample(time=resample_freq.pandas_freq).sum(dim="time")
     freq = check_freq(climate_vars[0].studied_data, dim="time")
-    return to_agg_units(result, climate_vars[0].studied_data, "count", deffreq=freq)
+    return _safe_to_agg_units(result, climate_vars[0].studied_data, "count", deffreq=freq)
 
 
 def excess(
@@ -301,7 +302,7 @@ def excess(
     )
     res = res.assign_attrs(units=f"delta_{res.attrs['units']}")
     freq = check_freq(study, dim="time")
-    return to_agg_units(res, study, "integral", deffreq=freq)
+    return _safe_to_agg_units(res, study, "integral", deffreq=freq)
 
 
 def deficit(
@@ -338,7 +339,7 @@ def deficit(
     res = deficit.clip(min=0).resample(time=resample_freq.pandas_freq).sum(dim="time")
     res = res.assign_attrs(units=f"delta_{res.attrs['units']}")
     freq = check_freq(study, dim="time")
-    return to_agg_units(res, study, "integral", deffreq=freq)
+    return _safe_to_agg_units(res, study, "integral", deffreq=freq)
 
 
 def fraction_of_total(
@@ -1572,3 +1573,12 @@ def check_freq(da: xr.DataArray, dim: str = "time", strict: bool = True) -> str 
             msg = "[icclim] Unable to infer frequency"
             raise ValueError(msg) from err
         return None
+
+def _safe_to_agg_units(*args, **kwargs) -> DataArray:
+    with warnings.catch_warnings():
+        warnings.filterwarnings(
+            "ignore",
+            category=UserWarning,
+            message=".*Unable to find the sampling frequency.*",
+        )
+        return to_agg_units(*args, **kwargs)